Fixed bug #26 first click on preview button
Now when the preview button is clicked for the first time, the
transition from editor.css will fire and the view will slide from right
to left, instead of just appearing.

Changed the editor.css, removed the padding:4px; from editor-preview class for the scroll-bar to be showed right.

Close #31